FEAT :sparkles: create architecture folder and move file
