+++ ambiance-back/src/app.module.ts
@@ -4,9 +4,10 @@ import { ConfigModule } from '@nestjs/config';
 import configuration from './config/configuration';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { PostsController } from './posts/posts.controller';
-import { PostsModule } from './posts/posts.module';
-
+import { PostsController } from './controllers/posts/posts.controller';
+import { PostsModule } from './modules/posts/posts.module';
+import { Post } from './entities/posts.entity';
+import { User } from './entities/users.entity';
 @Module({
   controllers: [AppController, PostsController],
   imports: [
@@ -21,7 +22,7 @@ import { PostsModule } from './posts/posts.module';
       username: process.env.DATABASE_USER,
       password: process.env.DATABASE_PASSWORD,
       database: process.env.DATABASE_NAME,
-      entities: [__dirname + '/**/*.entity{.ts,.js}'],
+      entities: [Post, User], // Ajouter les entités ici
       synchronize: false, // Permet de manipuler les entités de la base de données avec les fichiers entity.ts en temps réel
     }),
     PostsModule,
