+++ ambiance-back/src/entities/posts.entity.ts
@@ -5,7 +5,7 @@ import {
   ManyToOne,
   JoinColumn,
 } from 'typeorm';
-import { User } from '../users/users.entity';
+import { User } from './users.entity';
 
 @Entity('Publications')
 export class Post {
