+++ ambiance-back/src/modules/posts/posts.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { PostsService } from './posts.service';
-import { PostsController } from './posts.controller';
-import { Post } from './posts.entity';
+import { PostsService } from '../../services/posts/posts.service';
+import { PostsController } from '../../controllers/posts/posts.controller';
+import { Post } from '../../entities/posts.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Post])], // Définir l'entité Post ici
