+++ ambiance-back/src/services/posts/posts.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { Post } from './posts.entity';
+import { Post } from '../../entities/posts.entity';
 
 @Injectable()
 export class PostsService {
